Updating to Go1.19 with unix tag
Also makes it so that it can crosscompile without cgo support for darwin
