+++ .github/workflows/go.yml
@@ -8,7 +8,7 @@ jobs:
   build:
     runs-on: ${{ matrix.os }}
     env:
-      GO_VERSION: 1.16
+      GO_VERSION: 1.19
       GOARCH: amd64
     strategy:
       matrix:
