+++ screenshot.go
@@ -6,7 +6,7 @@ import (
 	"image"
 )
 
-// CaptureDisplay captures whole region of displayIndex'th display.
+// CaptureDisplay captures whole region of displayIndex'th display, starts at 0 for main display.
 func CaptureDisplay(displayIndex int) (*image.RGBA, error) {
 	rect := GetDisplayBounds(displayIndex)
 	return CaptureRect(rect)
