+++ screenshot_darwin.go
@@ -1,4 +1,4 @@
-// +build go1.10
+//go:build cgo
 
 package screenshot
 
@@ -11,7 +11,7 @@ void* CompatCGDisplayCreateImageForRect(CGDirectDisplayID display, CGRect rect) 
 }
 
 void CompatCGImageRelease(void* image) {
-	CGImageRelease(image);	
+	CGImageRelease(image);
 }
 
 void* CompatCGImageCreateCopyWithColorSpace(void* image, CGColorSpaceRef space) {
