--- screenshot_freebsd.go
@@ -1,1 +1,0 @@
-screenshot_linux.go
