--- screenshot_solaris.go
@@ -1,1 +1,0 @@
-screenshot_linux.go
